+++ assets/src/components/activities/check_all_that_apply/Settings.tsx
@@ -1,27 +1,21 @@
-import { isShuffled, toggleAnswerChoiceShuffling } from 'components/activities/common/utils';
-import { SettingsComponent } from 'components/activities/common/authoring/settings/main';
 import React from 'react';
 import { CheckAllThatApplyModelSchema } from 'components/activities/check_all_that_apply/schema';
-import { Actions } from 'components/activities/check_all_that_apply/actions';
+import { CATAActions } from 'components/activities/check_all_that_apply/actions';
 import { isTargetedCATA } from 'components/activities/check_all_that_apply/utils';
+import { ActivitySettings } from 'components/activities/common/authoring/settings/ActivitySettings';
+import { shuffleAnswerChoiceSetting } from 'components/activities/common/authoring/settings/activitySettingsActions';
+import { useAuthoringElementContext } from 'components/activities/AuthoringElement';
 
-interface Props {
-  dispatch: (action: any) => void;
-  model: CheckAllThatApplyModelSchema;
-}
-export const CheckAllThatApplySettings: React.FC<Props> = ({ dispatch, model }) => {
+export const CATASettingsConnected: React.FC = () => {
+  const { dispatch, model } = useAuthoringElementContext();
   return (
-    <SettingsComponent
+    <ActivitySettings
       settings={[
-        {
-          isEnabled: isShuffled(model.authoring.transformations),
-          label: 'Shuffle answer choice order',
-          onToggle: () => dispatch(toggleAnswerChoiceShuffling()),
-        },
+        shuffleAnswerChoiceSetting(model, dispatch),
         {
           isEnabled: isTargetedCATA(model),
           label: 'Targeted feedback',
-          onToggle: () => dispatch(Actions.toggleType()),
+          onToggle: () => dispatch(CATAActions.toggleType()),
         },
       ]}
     />