+++ assets/src/components/activities/common/authoring/settings/activitySettingsActions.ts
@@ -1,0 +1,11 @@
+import { isShuffled, toggleAnswerChoiceShuffling } from 'components/activities/common/utils';
+import { HasTransformations } from 'components/activities/types';
+
+export const shuffleAnswerChoiceSetting = (
+  model: HasTransformations,
+  dispatch: (action: any) => void,
+) => ({
+  isEnabled: isShuffled(model.authoring.transformations),
+  label: 'Shuffle answer choice order',
+  onToggle: () => dispatch(toggleAnswerChoiceShuffling()),
+});