+++ assets/src/components/parts/janus-Carousel/Carousel.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable react/prop-types */
 import React, { createRef, CSSProperties, useCallback, useEffect, useState } from 'react';
 import { Swiper, SwiperSlide } from 'swiper/react';
+import SwiperCore, { Navigation, Pagination, A11y, Keyboard, Zoom } from 'swiper';
 import { CapiVariableTypes } from '../../../adaptivity/capi';
 import {
   NotificationType,
@@ -20,7 +21,8 @@ const Carousel: React.FC<any> = (props) => {
   const [carouselCustomCss, setCarouselCustomCss] = useState<string>('');
   const [carouselZoom, setCarouselZoom] = useState<boolean>(true);
   const [cssClass, setCssClass] = useState('');
-
+  // initialize the swiper
+  SwiperCore.use([Navigation, Pagination, A11y, Keyboard, Zoom]);
   const initialize = useCallback(async (pModel) => {
     // set defaults
     const dZoom = typeof pModel.zoom === 'boolean' ? pModel.zoom : carouselZoom;
@@ -179,8 +181,7 @@ const Carousel: React.FC<any> = (props) => {
   } = model;
 
   const [carouselMode, setCarouselMode] = useState<string>(mode);
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const carouselDefaultCss = require('./Carousel.css');
+
   const MAGIC_NUMBER = 64;
   const PAGINATION_HEIGHT = 32;
   const styles: CSSProperties = {
@@ -274,12 +275,7 @@ const Carousel: React.FC<any> = (props) => {
       className={`janus-image-carousel ${cssClass}`}
       style={styles}
     >
-      {carouselCustomCss && (
-        <style type="text/css" style={{ display: 'none' }}>
-          {carouselDefaultCss}
-          {carouselCustomCss ? carouselCustomCss : null}
-        </style>
-      )}
+      {}
       {images.length > 0 && (
         <Swiper
           slidesPerView={1}
@@ -313,6 +309,6 @@ const Carousel: React.FC<any> = (props) => {
   ) : null;
 };
 
-export const tagName = 'janus-carousel';
+export const tagName = 'janus-image-carousel';
 
 export default Carousel;