+++ priv/repo/migrations/20210614145546_removeJanusCarouselPartRegistrations.exs
@@ -1,0 +1,10 @@
+defmodule Oli.Repo.Migrations.RemoveJanusCarousel do
+  use Ecto.Migration
+
+  def change do
+    execute "DELETE FROM part_component_registrations
+      WHERE delivery_element = 'janus-carousel';"
+
+      flush()
+  end
+end